Extract buffer structure processing to own class
